requirements.txt
certifi==2020.12.5
click==7.1.2
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
Werkzeug==1.0.1
WTForms==2.3.3
